requirements.txt
pypdf
pdfminer.six
regex
pandas
typer
pyyaml
